// src/app/api/model/aggregateCredits.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AggregateCrewMember } from './aggregateCrewMember';
import { AggregateCastMember } from './aggregateCastMember';


export interface AggregateCredits { 
    id?: number;
    cast?: Array<AggregateCastMember>;
    crew?: Array<AggregateCrewMember>;
}