// src/app/api/model/image.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Image { 
    aspect_ratio?: number;
    height?: number;
    iso_639_1?: string | null;
    file_path?: string;
    vote_average?: number;
    vote_count?: number;
    width?: number;
}